requirements.txt
python-dotenv~=1.0.0
SQLAlchemy~=2.0.23
pypdf2~=3.0.1
mock-alchemy~=0.2.6
